\name{anno}
\alias{anno}
\docType{data}
\title{
annotation profile of two sample data sets
}
\description{
The data give the functional annotation profiles of two sample data sets.
}
\usage{data(anno)}
\format{
  The format is:
List of 4
each element represents functional annotation profile of one sample
}
\details{
Metagenome ID 

obese     mgm4440613.3

lean      mgm4440823.3
}

\examples{
data(anno)
}
