\name{mgp10Anno}
\alias{mgp10Anno}
\docType{data}
\title{
KO abundance of metagenomic data
}
\description{
The data give the estimated KO abundance in four samples of mgp10 in MGRAST.
}
\usage{data(mgp10Anno)}
\format{
  The format is:
List of 4
each element represents KO abundance profile of one sample
}
\details{
Metagenome ID in MGRAST of this four sample

obese    mgm4440616.3  mgm4440613.3

lean     mgm4440824.3  mgm4440823.3
}

\examples{
data(mgp10Anno)
}
